requirements.txt
azureml-sdk==1.0.2
azure-cli==2.0.53
fnss==0.8.2